Proper version of valid_state_set_stack_ptr
I tried to prove separately a lemma for set_move and a lemma for
set_stack_ptr, but failed. So the proved lemma is about the composition
of set_move and set_stack_ptr, as before.

If we apply set_move then set_stack_ptr, it looks like the intermediate
rmap does not satisfy valid_state. If we apply set_stack_ptr then
set_move, then it looks wrong.

There should be some way to better unify the proofs. In the 3 cases
(stack, reg ptr, stack ptr), we call set_move, but never in the same
context, so it is not obvious.